Split silo.proto into core and admin services
Separate the gRPC API into two services to clarify which RPCs are
worker-facing vs operator-facing:

- `Silo` service (silo.proto): core worker operations — enqueue, lease,
  heartbeat, query, etc.
- `SiloAdmin` service (silo_admin.proto): admin/operator operations —
  splits, compaction, node info, profiling, imports, reset, etc.

Both services are registered on the same gRPC server and share the
same `silo.v1` package, so all message types remain in a single Rust
module. The split is purely at the service/trait level.

Also fixes the webui to fetch LSM state and trigger compaction via
ClusterClient (routing to remote nodes) instead of only working for
locally-owned shards.
