+++ .ai/AGENTS.md
@@ -60,7 +60,13 @@ Example configs are in `example_configs/`. The server reads TOML config files. S
 
 # Protocol Buffers
 
-Proto definitions are in `proto/silo.proto` and `proto/gubernator.proto`. Rust code is generated by `build.rs` during cargo build. If you modify proto files, just run `cargo build` - no separate protoc step needed.
+Proto definitions are split across two files in the same `silo.v1` package:
+
+- `proto/silo.proto` - Core worker-facing service (`Silo`): enqueue, lease, heartbeat, query, etc.
+- `proto/silo_admin.proto` - Admin/operator service (`SiloAdmin`): splits, compaction, node info, profiling, imports, etc. Used by the webui and siloctl.
+- `proto/gubernator.proto` - Rate limiting (Gubernator client).
+
+Rust code is generated by `build.rs` during cargo build. If you modify proto files, just run `cargo build` - no separate protoc step needed. When adding a new RPC, put it in the correct service based on whether it's worker-facing or operator-facing.
 
 # Rust Style
 
@@ -164,3 +170,7 @@ We use text-based sigils to ensure that the invariants and assertions in the All
 CI validates that all the sigils in the Alloy model are present in the Rust implementation. If changing the Alloy model to add new key invariants or assertions, complete your work on the Alloy model, then ensure you update the Rust code to make sure the sigils match. If the change to the Alloy model is large, ask if we should do the Rust work first, but if it is small, don't ask and just update the Rust code right away.
 
 Don't invent new sigils in the Rust code that don't have corresponding assertions in the Alloy models.
+
+# Web UI
+
+The webui is served by any node but must work correctly for **all** shards, not just locally-owned ones. Never access `ShardFactory` directly from a webui handler to read shard data — always go through `ClusterClient`, which routes to the owning node via gRPC when the shard is remote. If a new feature needs data from a `JobStoreShard`, add or extend a gRPC RPC so the cluster client can fetch it from any node, then use that cluster client method in the webui handler.
