+++ typescript_client/src/pb/silo_admin.client.ts
@@ -1,0 +1,246 @@
+// @generated by protobuf-ts 2.11.1 with parameter long_type_bigint,generate_dependencies
+// @generated from protobuf file "silo_admin.proto" (package "silo.v1", syntax proto3)
+// tslint:disable
+import type { RpcTransport } from "@protobuf-ts/runtime-rpc";
+import type { ServiceInfo } from "@protobuf-ts/runtime-rpc";
+import { SiloAdmin } from "./silo_admin";
+import type { ReadLsmStateResponse } from "./silo_admin";
+import type { ReadLsmStateRequest } from "./silo_admin";
+import type { CompactShardResponse } from "./silo_admin";
+import type { CompactShardRequest } from "./silo_admin";
+import type { ForceReleaseShardResponse } from "./silo_admin";
+import type { ForceReleaseShardRequest } from "./silo_admin";
+import type { ResetShardsResponse } from "./silo_admin";
+import type { ResetShardsRequest } from "./silo_admin";
+import type { ImportJobsResponse } from "./silo_admin";
+import type { ImportJobsRequest } from "./silo_admin";
+import type { ConfigureShardResponse } from "./silo_admin";
+import type { ConfigureShardRequest } from "./silo_admin";
+import type { GetSplitStatusResponse } from "./silo_admin";
+import type { GetSplitStatusRequest } from "./silo_admin";
+import type { RequestSplitResponse } from "./silo_admin";
+import type { RequestSplitRequest } from "./silo_admin";
+import type { CpuProfileResponse } from "./silo_admin";
+import type { CpuProfileRequest } from "./silo_admin";
+import { stackIntercept } from "@protobuf-ts/runtime-rpc";
+import type { GetNodeInfoResponse } from "./silo_admin";
+import type { GetNodeInfoRequest } from "./silo_admin";
+import type { UnaryCall } from "@protobuf-ts/runtime-rpc";
+import type { RpcOptions } from "@protobuf-ts/runtime-rpc";
+/**
+ * Admin service for operator/management operations.
+ * Used by the web UI, siloctl CLI, and monitoring tools.
+ * These RPCs are not part of the worker-facing API.
+ *
+ * @generated from protobuf service silo.v1.SiloAdmin
+ */
+export interface ISiloAdminClient {
+    /**
+     * Get information about this node and all the shards it owns.
+     *
+     * @generated from protobuf rpc: GetNodeInfo
+     */
+    getNodeInfo(input: GetNodeInfoRequest, options?: RpcOptions): UnaryCall<GetNodeInfoRequest, GetNodeInfoResponse>;
+    /**
+     * Capture a CPU profile from this node.
+     * Returns pprof protobuf data that can be analyzed with pprof or go tool pprof.
+     * The profile captures CPU usage for the specified duration.
+     *
+     * @generated from protobuf rpc: CpuProfile
+     */
+    cpuProfile(input: CpuProfileRequest, options?: RpcOptions): UnaryCall<CpuProfileRequest, CpuProfileResponse>;
+    /**
+     * Request a shard split operation.
+     * Initiates splitting a shard into two child shards at the specified split point.
+     * Returns FAILED_PRECONDITION if a split is already in progress.
+     * Returns NOT_FOUND if the shard doesn't exist on this node.
+     * Returns INVALID_ARGUMENT if the split point is outside the shard's range.
+     *
+     * @generated from protobuf rpc: RequestSplit
+     */
+    requestSplit(input: RequestSplitRequest, options?: RpcOptions): UnaryCall<RequestSplitRequest, RequestSplitResponse>;
+    /**
+     * Get the status of a shard split operation.
+     * Returns the current phase and child shard IDs if a split is in progress.
+     * If no split is in progress, returns with in_progress=false.
+     *
+     * @generated from protobuf rpc: GetSplitStatus
+     */
+    getSplitStatus(input: GetSplitStatusRequest, options?: RpcOptions): UnaryCall<GetSplitStatusRequest, GetSplitStatusResponse>;
+    /**
+     * Configure a shard's placement ring.
+     * Changes which placement ring the shard belongs to, affecting which nodes can own it.
+     * The shard will be handed off to a node that participates in the new ring.
+     * Returns the previous and current ring assignments.
+     *
+     * @generated from protobuf rpc: ConfigureShard
+     */
+    configureShard(input: ConfigureShardRequest, options?: RpcOptions): UnaryCall<ConfigureShardRequest, ConfigureShardResponse>;
+    /**
+     * Import jobs from another system with historical attempts.
+     * Unlike Enqueue, ImportJobs accepts completed attempt records and lets Silo
+     * take ownership going forward. Used for migrating workloads from other job queues.
+     * Each job is imported independently; per-job errors are returned in the response.
+     *
+     * @generated from protobuf rpc: ImportJobs
+     */
+    importJobs(input: ImportJobsRequest, options?: RpcOptions): UnaryCall<ImportJobsRequest, ImportJobsResponse>;
+    /**
+     * Reset all shards owned by this server.
+     * WARNING: Destructive operation. Only available in dev mode.
+     * Clears all jobs, tasks, queues, and other data.
+     *
+     * @generated from protobuf rpc: ResetShards
+     */
+    resetShards(input: ResetShardsRequest, options?: RpcOptions): UnaryCall<ResetShardsRequest, ResetShardsResponse>;
+    /**
+     * Force-release a shard lease regardless of the current holder.
+     * Operator escape hatch for recovering from permanently lost nodes.
+     * After force-release, any live node that desires the shard can acquire it.
+     *
+     * @generated from protobuf rpc: ForceReleaseShard
+     */
+    forceReleaseShard(input: ForceReleaseShardRequest, options?: RpcOptions): UnaryCall<ForceReleaseShardRequest, ForceReleaseShardResponse>;
+    /**
+     * Trigger a full compaction on a shard's SlateDB instance.
+     * Merges all L0 SSTs and sorted runs into a single sorted run, removing tombstones.
+     * This is useful for reclaiming space after bulk deletes or high job throughput.
+     *
+     * @generated from protobuf rpc: CompactShard
+     */
+    compactShard(input: CompactShardRequest, options?: RpcOptions): UnaryCall<CompactShardRequest, CompactShardResponse>;
+    /**
+     * Read the LSM tree state of a shard's SlateDB instance.
+     * Returns information about L0 SSTs and sorted runs to help operators
+     * understand storage health and whether compaction is needed.
+     *
+     * @generated from protobuf rpc: ReadLsmState
+     */
+    readLsmState(input: ReadLsmStateRequest, options?: RpcOptions): UnaryCall<ReadLsmStateRequest, ReadLsmStateResponse>;
+}
+/**
+ * Admin service for operator/management operations.
+ * Used by the web UI, siloctl CLI, and monitoring tools.
+ * These RPCs are not part of the worker-facing API.
+ *
+ * @generated from protobuf service silo.v1.SiloAdmin
+ */
+export class SiloAdminClient implements ISiloAdminClient, ServiceInfo {
+    typeName = SiloAdmin.typeName;
+    methods = SiloAdmin.methods;
+    options = SiloAdmin.options;
+    constructor(private readonly _transport: RpcTransport) {
+    }
+    /**
+     * Get information about this node and all the shards it owns.
+     *
+     * @generated from protobuf rpc: GetNodeInfo
+     */
+    getNodeInfo(input: GetNodeInfoRequest, options?: RpcOptions): UnaryCall<GetNodeInfoRequest, GetNodeInfoResponse> {
+        const method = this.methods[0], opt = this._transport.mergeOptions(options);
+        return stackIntercept<GetNodeInfoRequest, GetNodeInfoResponse>("unary", this._transport, method, opt, input);
+    }
+    /**
+     * Capture a CPU profile from this node.
+     * Returns pprof protobuf data that can be analyzed with pprof or go tool pprof.
+     * The profile captures CPU usage for the specified duration.
+     *
+     * @generated from protobuf rpc: CpuProfile
+     */
+    cpuProfile(input: CpuProfileRequest, options?: RpcOptions): UnaryCall<CpuProfileRequest, CpuProfileResponse> {
+        const method = this.methods[1], opt = this._transport.mergeOptions(options);
+        return stackIntercept<CpuProfileRequest, CpuProfileResponse>("unary", this._transport, method, opt, input);
+    }
+    /**
+     * Request a shard split operation.
+     * Initiates splitting a shard into two child shards at the specified split point.
+     * Returns FAILED_PRECONDITION if a split is already in progress.
+     * Returns NOT_FOUND if the shard doesn't exist on this node.
+     * Returns INVALID_ARGUMENT if the split point is outside the shard's range.
+     *
+     * @generated from protobuf rpc: RequestSplit
+     */
+    requestSplit(input: RequestSplitRequest, options?: RpcOptions): UnaryCall<RequestSplitRequest, RequestSplitResponse> {
+        const method = this.methods[2], opt = this._transport.mergeOptions(options);
+        return stackIntercept<RequestSplitRequest, RequestSplitResponse>("unary", this._transport, method, opt, input);
+    }
+    /**
+     * Get the status of a shard split operation.
+     * Returns the current phase and child shard IDs if a split is in progress.
+     * If no split is in progress, returns with in_progress=false.
+     *
+     * @generated from protobuf rpc: GetSplitStatus
+     */
+    getSplitStatus(input: GetSplitStatusRequest, options?: RpcOptions): UnaryCall<GetSplitStatusRequest, GetSplitStatusResponse> {
+        const method = this.methods[3], opt = this._transport.mergeOptions(options);
+        return stackIntercept<GetSplitStatusRequest, GetSplitStatusResponse>("unary", this._transport, method, opt, input);
+    }
+    /**
+     * Configure a shard's placement ring.
+     * Changes which placement ring the shard belongs to, affecting which nodes can own it.
+     * The shard will be handed off to a node that participates in the new ring.
+     * Returns the previous and current ring assignments.
+     *
+     * @generated from protobuf rpc: ConfigureShard
+     */
+    configureShard(input: ConfigureShardRequest, options?: RpcOptions): UnaryCall<ConfigureShardRequest, ConfigureShardResponse> {
+        const method = this.methods[4], opt = this._transport.mergeOptions(options);
+        return stackIntercept<ConfigureShardRequest, ConfigureShardResponse>("unary", this._transport, method, opt, input);
+    }
+    /**
+     * Import jobs from another system with historical attempts.
+     * Unlike Enqueue, ImportJobs accepts completed attempt records and lets Silo
+     * take ownership going forward. Used for migrating workloads from other job queues.
+     * Each job is imported independently; per-job errors are returned in the response.
+     *
+     * @generated from protobuf rpc: ImportJobs
+     */
+    importJobs(input: ImportJobsRequest, options?: RpcOptions): UnaryCall<ImportJobsRequest, ImportJobsResponse> {
+        const method = this.methods[5], opt = this._transport.mergeOptions(options);
+        return stackIntercept<ImportJobsRequest, ImportJobsResponse>("unary", this._transport, method, opt, input);
+    }
+    /**
+     * Reset all shards owned by this server.
+     * WARNING: Destructive operation. Only available in dev mode.
+     * Clears all jobs, tasks, queues, and other data.
+     *
+     * @generated from protobuf rpc: ResetShards
+     */
+    resetShards(input: ResetShardsRequest, options?: RpcOptions): UnaryCall<ResetShardsRequest, ResetShardsResponse> {
+        const method = this.methods[6], opt = this._transport.mergeOptions(options);
+        return stackIntercept<ResetShardsRequest, ResetShardsResponse>("unary", this._transport, method, opt, input);
+    }
+    /**
+     * Force-release a shard lease regardless of the current holder.
+     * Operator escape hatch for recovering from permanently lost nodes.
+     * After force-release, any live node that desires the shard can acquire it.
+     *
+     * @generated from protobuf rpc: ForceReleaseShard
+     */
+    forceReleaseShard(input: ForceReleaseShardRequest, options?: RpcOptions): UnaryCall<ForceReleaseShardRequest, ForceReleaseShardResponse> {
+        const method = this.methods[7], opt = this._transport.mergeOptions(options);
+        return stackIntercept<ForceReleaseShardRequest, ForceReleaseShardResponse>("unary", this._transport, method, opt, input);
+    }
+    /**
+     * Trigger a full compaction on a shard's SlateDB instance.
+     * Merges all L0 SSTs and sorted runs into a single sorted run, removing tombstones.
+     * This is useful for reclaiming space after bulk deletes or high job throughput.
+     *
+     * @generated from protobuf rpc: CompactShard
+     */
+    compactShard(input: CompactShardRequest, options?: RpcOptions): UnaryCall<CompactShardRequest, CompactShardResponse> {
+        const method = this.methods[8], opt = this._transport.mergeOptions(options);
+        return stackIntercept<CompactShardRequest, CompactShardResponse>("unary", this._transport, method, opt, input);
+    }
+    /**
+     * Read the LSM tree state of a shard's SlateDB instance.
+     * Returns information about L0 SSTs and sorted runs to help operators
+     * understand storage health and whether compaction is needed.
+     *
+     * @generated from protobuf rpc: ReadLsmState
+     */
+    readLsmState(input: ReadLsmStateRequest, options?: RpcOptions): UnaryCall<ReadLsmStateRequest, ReadLsmStateResponse> {
+        const method = this.methods[9], opt = this._transport.mergeOptions(options);
+        return stackIntercept<ReadLsmStateRequest, ReadLsmStateResponse>("unary", this._transport, method, opt, input);
+    }
+}
