+++ typescript_client/src/pb/silo_admin.ts
@@ -1,0 +1,2096 @@
+// @generated by protobuf-ts 2.11.1 with parameter long_type_bigint,generate_dependencies
+// @generated from protobuf file "silo_admin.proto" (package "silo.v1", syntax proto3)
+// tslint:disable
+import { ServiceType } from "@protobuf-ts/runtime-rpc";
+import { WireType } from "@protobuf-ts/runtime";
+import type { BinaryWriteOptions } from "@protobuf-ts/runtime";
+import type { IBinaryWriter } from "@protobuf-ts/runtime";
+import type { BinaryReadOptions } from "@protobuf-ts/runtime";
+import type { IBinaryReader } from "@protobuf-ts/runtime";
+import { UnknownFieldHandler } from "@protobuf-ts/runtime";
+import type { PartialMessage } from "@protobuf-ts/runtime";
+import { reflectionMergePartial } from "@protobuf-ts/runtime";
+import { MessageType } from "@protobuf-ts/runtime";
+import { JobStatus } from "./silo";
+import { Limit } from "./silo";
+import { RetryPolicy } from "./silo";
+import { SerializedBytes } from "./silo";
+import { AttemptStatus } from "./silo";
+/**
+ * Request to reset all shards owned by this server.
+ * WARNING: Destructive operation. Only available in dev mode.
+ *
+ * @generated from protobuf message silo.v1.ResetShardsRequest
+ */
+export interface ResetShardsRequest {
+}
+/**
+ * Response confirming shards were reset.
+ *
+ * @generated from protobuf message silo.v1.ResetShardsResponse
+ */
+export interface ResetShardsResponse {
+    /**
+     * @generated from protobuf field: uint32 shards_reset = 1
+     */
+    shardsReset: number; // Number of shards that were cleared.
+}
+/**
+ * Request to capture a CPU profile from this node.
+ * Used for production debugging and performance analysis.
+ *
+ * @generated from protobuf message silo.v1.CpuProfileRequest
+ */
+export interface CpuProfileRequest {
+    /**
+     * @generated from protobuf field: uint32 duration_seconds = 1
+     */
+    durationSeconds: number; // How long to profile (1-300 seconds). Default 30.
+    /**
+     * @generated from protobuf field: uint32 frequency = 2
+     */
+    frequency: number; // Sampling frequency in Hz (1-1000). Default 100.
+}
+/**
+ * CPU profile data in pprof protobuf format.
+ * Can be analyzed with `pprof` or `go tool pprof`.
+ *
+ * @generated from protobuf message silo.v1.CpuProfileResponse
+ */
+export interface CpuProfileResponse {
+    /**
+     * @generated from protobuf field: bytes profile_data = 1
+     */
+    profileData: Uint8Array; // pprof protobuf bytes (not gzip compressed).
+    /**
+     * @generated from protobuf field: uint32 duration_seconds = 2
+     */
+    durationSeconds: number; // Actual duration profiled.
+    /**
+     * @generated from protobuf field: uint64 samples = 3
+     */
+    samples: bigint; // Number of samples collected.
+}
+/**
+ * Request to initiate a shard split operation.
+ *
+ * @generated from protobuf message silo.v1.RequestSplitRequest
+ */
+export interface RequestSplitRequest {
+    /**
+     * @generated from protobuf field: string shard_id = 1
+     */
+    shardId: string; // Shard ID (UUID) of the shard to split.
+    /**
+     * @generated from protobuf field: string split_point = 2
+     */
+    splitPoint: string; // Tenant ID where to split the keyspace.
+}
+/**
+ * Response after initiating a shard split.
+ *
+ * @generated from protobuf message silo.v1.RequestSplitResponse
+ */
+export interface RequestSplitResponse {
+    /**
+     * @generated from protobuf field: string left_child_id = 1
+     */
+    leftChildId: string; // UUID of the left child shard [parent_start, split_point).
+    /**
+     * @generated from protobuf field: string right_child_id = 2
+     */
+    rightChildId: string; // UUID of the right child shard [split_point, parent_end).
+    /**
+     * @generated from protobuf field: string phase = 3
+     */
+    phase: string; // Current split phase (e.g., "SplitRequested").
+}
+/**
+ * Request to get the status of a shard split operation.
+ *
+ * @generated from protobuf message silo.v1.GetSplitStatusRequest
+ */
+export interface GetSplitStatusRequest {
+    /**
+     * @generated from protobuf field: string shard_id = 1
+     */
+    shardId: string; // Parent shard ID (UUID) of the split operation.
+}
+/**
+ * Response with the current split status.
+ * Returns empty if no split is in progress for the shard.
+ *
+ * @generated from protobuf message silo.v1.GetSplitStatusResponse
+ */
+export interface GetSplitStatusResponse {
+    /**
+     * @generated from protobuf field: bool in_progress = 1
+     */
+    inProgress: boolean; // True if a split is in progress for this shard.
+    /**
+     * @generated from protobuf field: string phase = 2
+     */
+    phase: string; // Current split phase (empty if not in progress).
+    /**
+     * @generated from protobuf field: string left_child_id = 3
+     */
+    leftChildId: string; // UUID of the left child shard (empty if not in progress).
+    /**
+     * @generated from protobuf field: string right_child_id = 4
+     */
+    rightChildId: string; // UUID of the right child shard (empty if not in progress).
+    /**
+     * @generated from protobuf field: string split_point = 5
+     */
+    splitPoint: string; // Tenant ID at which the split occurs (empty if not in progress).
+    /**
+     * @generated from protobuf field: string initiator_node_id = 6
+     */
+    initiatorNodeId: string; // Node ID that initiated the split.
+    /**
+     * @generated from protobuf field: int64 requested_at_ms = 7
+     */
+    requestedAtMs: bigint; // Unix timestamp (ms) when split was requested.
+}
+/**
+ * Information about a shard owned by a node, including counters and cleanup status.
+ *
+ * @generated from protobuf message silo.v1.OwnedShardInfo
+ */
+export interface OwnedShardInfo {
+    /**
+     * @generated from protobuf field: string shard_id = 1
+     */
+    shardId: string; // The shard ID (UUID).
+    /**
+     * @generated from protobuf field: int64 total_jobs = 2
+     */
+    totalJobs: bigint; // Total number of jobs in the shard (not deleted).
+    /**
+     * @generated from protobuf field: int64 completed_jobs = 3
+     */
+    completedJobs: bigint; // Number of jobs in terminal states (Succeeded, Failed, Cancelled).
+    /**
+     * @generated from protobuf field: string cleanup_status = 4
+     */
+    cleanupStatus: string; // Cleanup status: "CompactionDone", "CleanupPending", "CleanupRunning", "CleanupDone".
+    /**
+     * @generated from protobuf field: int64 created_at_ms = 5
+     */
+    createdAtMs: bigint; // Unix timestamp (ms) when this shard was first created/initialized.
+    /**
+     * @generated from protobuf field: int64 cleanup_completed_at_ms = 6
+     */
+    cleanupCompletedAtMs: bigint; // Unix timestamp (ms) when cleanup completed (0 if not applicable or not completed).
+}
+/**
+ * Request to get node information including owned shards with their counters and cleanup status.
+ *
+ * @generated from protobuf message silo.v1.GetNodeInfoRequest
+ */
+export interface GetNodeInfoRequest {
+}
+/**
+ * Response with node information and details for all shards owned by this node.
+ *
+ * @generated from protobuf message silo.v1.GetNodeInfoResponse
+ */
+export interface GetNodeInfoResponse {
+    /**
+     * @generated from protobuf field: string node_id = 1
+     */
+    nodeId: string; // Unique identifier of this node.
+    /**
+     * @generated from protobuf field: repeated silo.v1.OwnedShardInfo owned_shards = 2
+     */
+    ownedShards: OwnedShardInfo[]; // Information for each shard owned by this node.
+    /**
+     * @generated from protobuf field: repeated string placement_rings = 3
+     */
+    placementRings: string[]; // Placement rings this node participates in.
+}
+/**
+ * Request to configure a shard's placement ring.
+ *
+ * @generated from protobuf message silo.v1.ConfigureShardRequest
+ */
+export interface ConfigureShardRequest {
+    /**
+     * @generated from protobuf field: string shard = 1
+     */
+    shard: string; // The shard ID (UUID) to configure.
+    /**
+     * @generated from protobuf field: optional string placement_ring = 2
+     */
+    placementRing?: string; // The placement ring to assign (empty/null = default ring).
+    /**
+     * @generated from protobuf field: optional string tenant = 100
+     */
+    tenant?: string; // Optional tenant ID (for multi-tenant mode).
+}
+/**
+ * Response after configuring a shard's placement ring.
+ *
+ * @generated from protobuf message silo.v1.ConfigureShardResponse
+ */
+export interface ConfigureShardResponse {
+    /**
+     * @generated from protobuf field: string previous_ring = 1
+     */
+    previousRing: string; // The placement ring before the change (empty = default ring).
+    /**
+     * @generated from protobuf field: string current_ring = 2
+     */
+    currentRing: string; // The placement ring after the change (empty = default ring).
+}
+/**
+ * A historical attempt record for job import.
+ * All attempts must be in terminal states (no Running).
+ *
+ * @generated from protobuf message silo.v1.ImportAttempt
+ */
+export interface ImportAttempt {
+    /**
+     * @generated from protobuf field: silo.v1.AttemptStatus status = 1
+     */
+    status: AttemptStatus; // Must be terminal (SUCCEEDED, FAILED, or CANCELLED).
+    /**
+     * @generated from protobuf field: int64 started_at_ms = 2
+     */
+    startedAtMs: bigint; // When the attempt started (epoch ms).
+    /**
+     * @generated from protobuf field: int64 finished_at_ms = 3
+     */
+    finishedAtMs: bigint; // When the attempt finished (epoch ms).
+    /**
+     * @generated from protobuf field: optional silo.v1.SerializedBytes result = 4
+     */
+    result?: SerializedBytes; // Present if succeeded.
+    /**
+     * @generated from protobuf field: optional string error_code = 5
+     */
+    errorCode?: string; // Present if failed.
+    /**
+     * @generated from protobuf field: optional silo.v1.SerializedBytes error_data = 6
+     */
+    errorData?: SerializedBytes; // Present if failed.
+}
+/**
+ * Request to import a single job from another system.
+ * Unlike Enqueue, ImportJob accepts historical attempts and lets Silo take ownership going forward.
+ *
+ * @generated from protobuf message silo.v1.ImportJobRequest
+ */
+export interface ImportJobRequest {
+    /**
+     * @generated from protobuf field: string shard = 1
+     */
+    shard: string; // Shard ID (UUID) where the job should be stored.
+    /**
+     * @generated from protobuf field: string id = 2
+     */
+    id: string; // Required job ID (migration preserves IDs).
+    /**
+     * @generated from protobuf field: uint32 priority = 3
+     */
+    priority: number; // Priority from 0-99. Lower is higher priority.
+    /**
+     * @generated from protobuf field: int64 enqueue_time_ms = 4
+     */
+    enqueueTimeMs: bigint; // Original enqueue time from source system (0 = now).
+    /**
+     * @generated from protobuf field: int64 start_at_ms = 5
+     */
+    startAtMs: bigint; // When the next attempt should start (0 = now, only for non-terminal).
+    /**
+     * @generated from protobuf field: optional silo.v1.RetryPolicy retry_policy = 6
+     */
+    retryPolicy?: RetryPolicy; // Retry configuration.
+    /**
+     * @generated from protobuf field: silo.v1.SerializedBytes payload = 7
+     */
+    payload?: SerializedBytes; // Opaque serialized payload.
+    /**
+     * @generated from protobuf field: repeated silo.v1.Limit limits = 8
+     */
+    limits: Limit[]; // Ordered list of limits.
+    /**
+     * @generated from protobuf field: optional string tenant = 9
+     */
+    tenant?: string; // Tenant ID for multi-tenant deployments.
+    /**
+     * @generated from protobuf field: map<string, string> metadata = 10
+     */
+    metadata: {
+        [key: string]: string;
+    }; // Arbitrary key/value metadata.
+    /**
+     * @generated from protobuf field: string task_group = 11
+     */
+    taskGroup: string; // Task group for organizing tasks.
+    /**
+     * @generated from protobuf field: repeated silo.v1.ImportAttempt attempts = 12
+     */
+    attempts: ImportAttempt[]; // Historical attempts, all terminal.
+}
+/**
+ * Batch request to import multiple jobs.
+ *
+ * @generated from protobuf message silo.v1.ImportJobsRequest
+ */
+export interface ImportJobsRequest {
+    /**
+     * @generated from protobuf field: repeated silo.v1.ImportJobRequest jobs = 1
+     */
+    jobs: ImportJobRequest[];
+}
+/**
+ * Result of importing a single job.
+ *
+ * @generated from protobuf message silo.v1.ImportJobResult
+ */
+export interface ImportJobResult {
+    /**
+     * @generated from protobuf field: string id = 1
+     */
+    id: string; // The job's ID.
+    /**
+     * @generated from protobuf field: bool success = 2
+     */
+    success: boolean; // Whether the import succeeded.
+    /**
+     * @generated from protobuf field: optional string error = 3
+     */
+    error?: string; // Error message if import failed.
+    /**
+     * @generated from protobuf field: silo.v1.JobStatus status = 4
+     */
+    status: JobStatus; // The determined status of the imported job.
+}
+/**
+ * Response containing results for each imported job.
+ *
+ * @generated from protobuf message silo.v1.ImportJobsResponse
+ */
+export interface ImportJobsResponse {
+    /**
+     * @generated from protobuf field: repeated silo.v1.ImportJobResult results = 1
+     */
+    results: ImportJobResult[];
+}
+/**
+ * Request to force-release a shard's ownership lease.
+ * Operator escape hatch for permanently lost nodes.
+ *
+ * @generated from protobuf message silo.v1.ForceReleaseShardRequest
+ */
+export interface ForceReleaseShardRequest {
+    /**
+     * @generated from protobuf field: string shard = 1
+     */
+    shard: string; // The shard ID (UUID) to force-release.
+}
+/**
+ * Response after force-releasing a shard lease.
+ *
+ * @generated from protobuf message silo.v1.ForceReleaseShardResponse
+ */
+export interface ForceReleaseShardResponse {
+    /**
+     * @generated from protobuf field: bool released = 1
+     */
+    released: boolean; // True if the lease was released.
+}
+/**
+ * Request to trigger a full compaction on a shard's SlateDB instance.
+ * Merges all L0 SSTs and sorted runs into a single sorted run, removing tombstones.
+ *
+ * @generated from protobuf message silo.v1.CompactShardRequest
+ */
+export interface CompactShardRequest {
+    /**
+     * @generated from protobuf field: string shard = 1
+     */
+    shard: string; // The shard ID (UUID) to compact.
+}
+/**
+ * Response after submitting a full compaction.
+ *
+ * @generated from protobuf message silo.v1.CompactShardResponse
+ */
+export interface CompactShardResponse {
+    /**
+     * @generated from protobuf field: string status = 1
+     */
+    status: string; // Status message describing the compaction result.
+}
+/**
+ * Request to read the LSM tree state of a shard's SlateDB instance.
+ *
+ * @generated from protobuf message silo.v1.ReadLsmStateRequest
+ */
+export interface ReadLsmStateRequest {
+    /**
+     * @generated from protobuf field: string shard = 1
+     */
+    shard: string; // The shard ID (UUID) to read LSM state for.
+}
+/**
+ * Information about a single SST file in the L0 level.
+ *
+ * @generated from protobuf message silo.v1.LsmSstInfo
+ */
+export interface LsmSstInfo {
+    /**
+     * @generated from protobuf field: string id = 1
+     */
+    id: string; // SST identifier.
+    /**
+     * @generated from protobuf field: uint64 estimated_size = 2
+     */
+    estimatedSize: bigint; // Estimated size in bytes.
+}
+/**
+ * Information about a sorted run (compacted level).
+ *
+ * @generated from protobuf message silo.v1.LsmSortedRunInfo
+ */
+export interface LsmSortedRunInfo {
+    /**
+     * @generated from protobuf field: uint32 id = 1
+     */
+    id: number; // Sorted run identifier.
+    /**
+     * @generated from protobuf field: uint32 sst_count = 2
+     */
+    sstCount: number; // Number of SSTs in this sorted run.
+    /**
+     * @generated from protobuf field: uint64 estimated_size = 3
+     */
+    estimatedSize: bigint; // Estimated total size in bytes.
+}
+/**
+ * Response with the LSM tree state of a shard.
+ *
+ * @generated from protobuf message silo.v1.ReadLsmStateResponse
+ */
+export interface ReadLsmStateResponse {
+    /**
+     * @generated from protobuf field: repeated silo.v1.LsmSstInfo l0_ssts = 1
+     */
+    l0Ssts: LsmSstInfo[]; // L0 SSTs (unflushed/uncompacted).
+    /**
+     * @generated from protobuf field: repeated silo.v1.LsmSortedRunInfo sorted_runs = 2
+     */
+    sortedRuns: LsmSortedRunInfo[]; // Sorted runs (compacted).
+    /**
+     * @generated from protobuf field: uint64 total_l0_size = 3
+     */
+    totalL0Size: bigint; // Total estimated size of all L0 SSTs in bytes.
+    /**
+     * @generated from protobuf field: uint64 total_sorted_run_size = 4
+     */
+    totalSortedRunSize: bigint; // Total estimated size of all sorted runs in bytes.
+}
+// @generated message type with reflection information, may provide speed optimized methods
+class ResetShardsRequest$Type extends MessageType<ResetShardsRequest> {
+    constructor() {
+        super("silo.v1.ResetShardsRequest", []);
+    }
+    create(value?: PartialMessage<ResetShardsRequest>): ResetShardsRequest {
+        const message = globalThis.Object.create((this.messagePrototype!));
+        if (value !== undefined)
+            reflectionMergePartial<ResetShardsRequest>(this, message, value);
+        return message;
+    }
+    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: ResetShardsRequest): ResetShardsRequest {
+        let message = target ?? this.create(), end = reader.pos + length;
+        while (reader.pos < end) {
+            let [fieldNo, wireType] = reader.tag();
+            switch (fieldNo) {
+                default:
+                    let u = options.readUnknownField;
+                    if (u === "throw")
+                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
+                    let d = reader.skip(wireType);
+                    if (u !== false)
+                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
+            }
+        }
+        return message;
+    }
+    internalBinaryWrite(message: ResetShardsRequest, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
+        let u = options.writeUnknownFields;
+        if (u !== false)
+            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
+        return writer;
+    }
+}
+/**
+ * @generated MessageType for protobuf message silo.v1.ResetShardsRequest
+ */
+export const ResetShardsRequest = new ResetShardsRequest$Type();
+// @generated message type with reflection information, may provide speed optimized methods
+class ResetShardsResponse$Type extends MessageType<ResetShardsResponse> {
+    constructor() {
+        super("silo.v1.ResetShardsResponse", [
+            { no: 1, name: "shards_reset", kind: "scalar", T: 13 /*ScalarType.UINT32*/ }
+        ]);
+    }
+    create(value?: PartialMessage<ResetShardsResponse>): ResetShardsResponse {
+        const message = globalThis.Object.create((this.messagePrototype!));
+        message.shardsReset = 0;
+        if (value !== undefined)
+            reflectionMergePartial<ResetShardsResponse>(this, message, value);
+        return message;
+    }
+    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: ResetShardsResponse): ResetShardsResponse {
+        let message = target ?? this.create(), end = reader.pos + length;
+        while (reader.pos < end) {
+            let [fieldNo, wireType] = reader.tag();
+            switch (fieldNo) {
+                case /* uint32 shards_reset */ 1:
+                    message.shardsReset = reader.uint32();
+                    break;
+                default:
+                    let u = options.readUnknownField;
+                    if (u === "throw")
+                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
+                    let d = reader.skip(wireType);
+                    if (u !== false)
+                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
+            }
+        }
+        return message;
+    }
+    internalBinaryWrite(message: ResetShardsResponse, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
+        /* uint32 shards_reset = 1; */
+        if (message.shardsReset !== 0)
+            writer.tag(1, WireType.Varint).uint32(message.shardsReset);
+        let u = options.writeUnknownFields;
+        if (u !== false)
+            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
+        return writer;
+    }
+}
+/**
+ * @generated MessageType for protobuf message silo.v1.ResetShardsResponse
+ */
+export const ResetShardsResponse = new ResetShardsResponse$Type();
+// @generated message type with reflection information, may provide speed optimized methods
+class CpuProfileRequest$Type extends MessageType<CpuProfileRequest> {
+    constructor() {
+        super("silo.v1.CpuProfileRequest", [
+            { no: 1, name: "duration_seconds", kind: "scalar", T: 13 /*ScalarType.UINT32*/ },
+            { no: 2, name: "frequency", kind: "scalar", T: 13 /*ScalarType.UINT32*/ }
+        ]);
+    }
+    create(value?: PartialMessage<CpuProfileRequest>): CpuProfileRequest {
+        const message = globalThis.Object.create((this.messagePrototype!));
+        message.durationSeconds = 0;
+        message.frequency = 0;
+        if (value !== undefined)
+            reflectionMergePartial<CpuProfileRequest>(this, message, value);
+        return message;
+    }
+    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: CpuProfileRequest): CpuProfileRequest {
+        let message = target ?? this.create(), end = reader.pos + length;
+        while (reader.pos < end) {
+            let [fieldNo, wireType] = reader.tag();
+            switch (fieldNo) {
+                case /* uint32 duration_seconds */ 1:
+                    message.durationSeconds = reader.uint32();
+                    break;
+                case /* uint32 frequency */ 2:
+                    message.frequency = reader.uint32();
+                    break;
+                default:
+                    let u = options.readUnknownField;
+                    if (u === "throw")
+                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
+                    let d = reader.skip(wireType);
+                    if (u !== false)
+                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
+            }
+        }
+        return message;
+    }
+    internalBinaryWrite(message: CpuProfileRequest, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
+        /* uint32 duration_seconds = 1; */
+        if (message.durationSeconds !== 0)
+            writer.tag(1, WireType.Varint).uint32(message.durationSeconds);
+        /* uint32 frequency = 2; */
+        if (message.frequency !== 0)
+            writer.tag(2, WireType.Varint).uint32(message.frequency);
+        let u = options.writeUnknownFields;
+        if (u !== false)
+            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
+        return writer;
+    }
+}
+/**
+ * @generated MessageType for protobuf message silo.v1.CpuProfileRequest
+ */
+export const CpuProfileRequest = new CpuProfileRequest$Type();
+// @generated message type with reflection information, may provide speed optimized methods
+class CpuProfileResponse$Type extends MessageType<CpuProfileResponse> {
+    constructor() {
+        super("silo.v1.CpuProfileResponse", [
+            { no: 1, name: "profile_data", kind: "scalar", T: 12 /*ScalarType.BYTES*/ },
+            { no: 2, name: "duration_seconds", kind: "scalar", T: 13 /*ScalarType.UINT32*/ },
+            { no: 3, name: "samples", kind: "scalar", T: 4 /*ScalarType.UINT64*/, L: 0 /*LongType.BIGINT*/ }
+        ]);
+    }
+    create(value?: PartialMessage<CpuProfileResponse>): CpuProfileResponse {
+        const message = globalThis.Object.create((this.messagePrototype!));
+        message.profileData = new Uint8Array(0);
+        message.durationSeconds = 0;
+        message.samples = 0n;
+        if (value !== undefined)
+            reflectionMergePartial<CpuProfileResponse>(this, message, value);
+        return message;
+    }
+    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: CpuProfileResponse): CpuProfileResponse {
+        let message = target ?? this.create(), end = reader.pos + length;
+        while (reader.pos < end) {
+            let [fieldNo, wireType] = reader.tag();
+            switch (fieldNo) {
+                case /* bytes profile_data */ 1:
+                    message.profileData = reader.bytes();
+                    break;
+                case /* uint32 duration_seconds */ 2:
+                    message.durationSeconds = reader.uint32();
+                    break;
+                case /* uint64 samples */ 3:
+                    message.samples = reader.uint64().toBigInt();
+                    break;
+                default:
+                    let u = options.readUnknownField;
+                    if (u === "throw")
+                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
+                    let d = reader.skip(wireType);
+                    if (u !== false)
+                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
+            }
+        }
+        return message;
+    }
+    internalBinaryWrite(message: CpuProfileResponse, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
+        /* bytes profile_data = 1; */
+        if (message.profileData.length)
+            writer.tag(1, WireType.LengthDelimited).bytes(message.profileData);
+        /* uint32 duration_seconds = 2; */
+        if (message.durationSeconds !== 0)
+            writer.tag(2, WireType.Varint).uint32(message.durationSeconds);
+        /* uint64 samples = 3; */
+        if (message.samples !== 0n)
+            writer.tag(3, WireType.Varint).uint64(message.samples);
+        let u = options.writeUnknownFields;
+        if (u !== false)
+            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
+        return writer;
+    }
+}
+/**
+ * @generated MessageType for protobuf message silo.v1.CpuProfileResponse
+ */
+export const CpuProfileResponse = new CpuProfileResponse$Type();
+// @generated message type with reflection information, may provide speed optimized methods
+class RequestSplitRequest$Type extends MessageType<RequestSplitRequest> {
+    constructor() {
+        super("silo.v1.RequestSplitRequest", [
+            { no: 1, name: "shard_id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
+            { no: 2, name: "split_point", kind: "scalar", T: 9 /*ScalarType.STRING*/ }
+        ]);
+    }
+    create(value?: PartialMessage<RequestSplitRequest>): RequestSplitRequest {
+        const message = globalThis.Object.create((this.messagePrototype!));
+        message.shardId = "";
+        message.splitPoint = "";
+        if (value !== undefined)
+            reflectionMergePartial<RequestSplitRequest>(this, message, value);
+        return message;
+    }
+    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: RequestSplitRequest): RequestSplitRequest {
+        let message = target ?? this.create(), end = reader.pos + length;
+        while (reader.pos < end) {
+            let [fieldNo, wireType] = reader.tag();
+            switch (fieldNo) {
+                case /* string shard_id */ 1:
+                    message.shardId = reader.string();
+                    break;
+                case /* string split_point */ 2:
+                    message.splitPoint = reader.string();
+                    break;
+                default:
+                    let u = options.readUnknownField;
+                    if (u === "throw")
+                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
+                    let d = reader.skip(wireType);
+                    if (u !== false)
+                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
+            }
+        }
+        return message;
+    }
+    internalBinaryWrite(message: RequestSplitRequest, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
+        /* string shard_id = 1; */
+        if (message.shardId !== "")
+            writer.tag(1, WireType.LengthDelimited).string(message.shardId);
+        /* string split_point = 2; */
+        if (message.splitPoint !== "")
+            writer.tag(2, WireType.LengthDelimited).string(message.splitPoint);
+        let u = options.writeUnknownFields;
+        if (u !== false)
+            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
+        return writer;
+    }
+}
+/**
+ * @generated MessageType for protobuf message silo.v1.RequestSplitRequest
+ */
+export const RequestSplitRequest = new RequestSplitRequest$Type();
+// @generated message type with reflection information, may provide speed optimized methods
+class RequestSplitResponse$Type extends MessageType<RequestSplitResponse> {
+    constructor() {
+        super("silo.v1.RequestSplitResponse", [
+            { no: 1, name: "left_child_id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
+            { no: 2, name: "right_child_id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
+            { no: 3, name: "phase", kind: "scalar", T: 9 /*ScalarType.STRING*/ }
+        ]);
+    }
+    create(value?: PartialMessage<RequestSplitResponse>): RequestSplitResponse {
+        const message = globalThis.Object.create((this.messagePrototype!));
+        message.leftChildId = "";
+        message.rightChildId = "";
+        message.phase = "";
+        if (value !== undefined)
+            reflectionMergePartial<RequestSplitResponse>(this, message, value);
+        return message;
+    }
+    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: RequestSplitResponse): RequestSplitResponse {
+        let message = target ?? this.create(), end = reader.pos + length;
+        while (reader.pos < end) {
+            let [fieldNo, wireType] = reader.tag();
+            switch (fieldNo) {
+                case /* string left_child_id */ 1:
+                    message.leftChildId = reader.string();
+                    break;
+                case /* string right_child_id */ 2:
+                    message.rightChildId = reader.string();
+                    break;
+                case /* string phase */ 3:
+                    message.phase = reader.string();
+                    break;
+                default:
+                    let u = options.readUnknownField;
+                    if (u === "throw")
+                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
+                    let d = reader.skip(wireType);
+                    if (u !== false)
+                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
+            }
+        }
+        return message;
+    }
+    internalBinaryWrite(message: RequestSplitResponse, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
+        /* string left_child_id = 1; */
+        if (message.leftChildId !== "")
+            writer.tag(1, WireType.LengthDelimited).string(message.leftChildId);
+        /* string right_child_id = 2; */
+        if (message.rightChildId !== "")
+            writer.tag(2, WireType.LengthDelimited).string(message.rightChildId);
+        /* string phase = 3; */
+        if (message.phase !== "")
+            writer.tag(3, WireType.LengthDelimited).string(message.phase);
+        let u = options.writeUnknownFields;
+        if (u !== false)
+            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
+        return writer;
+    }
+}
+/**
+ * @generated MessageType for protobuf message silo.v1.RequestSplitResponse
+ */
+export const RequestSplitResponse = new RequestSplitResponse$Type();
+// @generated message type with reflection information, may provide speed optimized methods
+class GetSplitStatusRequest$Type extends MessageType<GetSplitStatusRequest> {
+    constructor() {
+        super("silo.v1.GetSplitStatusRequest", [
+            { no: 1, name: "shard_id", kind: "scalar", T: 9 /*ScalarType.STRING*/ }
+        ]);
+    }
+    create(value?: PartialMessage<GetSplitStatusRequest>): GetSplitStatusRequest {
+        const message = globalThis.Object.create((this.messagePrototype!));
+        message.shardId = "";
+        if (value !== undefined)
+            reflectionMergePartial<GetSplitStatusRequest>(this, message, value);
+        return message;
+    }
+    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: GetSplitStatusRequest): GetSplitStatusRequest {
+        let message = target ?? this.create(), end = reader.pos + length;
+        while (reader.pos < end) {
+            let [fieldNo, wireType] = reader.tag();
+            switch (fieldNo) {
+                case /* string shard_id */ 1:
+                    message.shardId = reader.string();
+                    break;
+                default:
+                    let u = options.readUnknownField;
+                    if (u === "throw")
+                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
+                    let d = reader.skip(wireType);
+                    if (u !== false)
+                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
+            }
+        }
+        return message;
+    }
+    internalBinaryWrite(message: GetSplitStatusRequest, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
+        /* string shard_id = 1; */
+        if (message.shardId !== "")
+            writer.tag(1, WireType.LengthDelimited).string(message.shardId);
+        let u = options.writeUnknownFields;
+        if (u !== false)
+            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
+        return writer;
+    }
+}
+/**
+ * @generated MessageType for protobuf message silo.v1.GetSplitStatusRequest
+ */
+export const GetSplitStatusRequest = new GetSplitStatusRequest$Type();
+// @generated message type with reflection information, may provide speed optimized methods
+class GetSplitStatusResponse$Type extends MessageType<GetSplitStatusResponse> {
+    constructor() {
+        super("silo.v1.GetSplitStatusResponse", [
+            { no: 1, name: "in_progress", kind: "scalar", T: 8 /*ScalarType.BOOL*/ },
+            { no: 2, name: "phase", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
+            { no: 3, name: "left_child_id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
+            { no: 4, name: "right_child_id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
+            { no: 5, name: "split_point", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
+            { no: 6, name: "initiator_node_id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
+            { no: 7, name: "requested_at_ms", kind: "scalar", T: 3 /*ScalarType.INT64*/, L: 0 /*LongType.BIGINT*/ }
+        ]);
+    }
+    create(value?: PartialMessage<GetSplitStatusResponse>): GetSplitStatusResponse {
+        const message = globalThis.Object.create((this.messagePrototype!));
+        message.inProgress = false;
+        message.phase = "";
+        message.leftChildId = "";
+        message.rightChildId = "";
+        message.splitPoint = "";
+        message.initiatorNodeId = "";
+        message.requestedAtMs = 0n;
+        if (value !== undefined)
+            reflectionMergePartial<GetSplitStatusResponse>(this, message, value);
+        return message;
+    }
+    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: GetSplitStatusResponse): GetSplitStatusResponse {
+        let message = target ?? this.create(), end = reader.pos + length;
+        while (reader.pos < end) {
+            let [fieldNo, wireType] = reader.tag();
+            switch (fieldNo) {
+                case /* bool in_progress */ 1:
+                    message.inProgress = reader.bool();
+                    break;
+                case /* string phase */ 2:
+                    message.phase = reader.string();
+                    break;
+                case /* string left_child_id */ 3:
+                    message.leftChildId = reader.string();
+                    break;
+                case /* string right_child_id */ 4:
+                    message.rightChildId = reader.string();
+                    break;
+                case /* string split_point */ 5:
+                    message.splitPoint = reader.string();
+                    break;
+                case /* string initiator_node_id */ 6:
+                    message.initiatorNodeId = reader.string();
+                    break;
+                case /* int64 requested_at_ms */ 7:
+                    message.requestedAtMs = reader.int64().toBigInt();
+                    break;
+                default:
+                    let u = options.readUnknownField;
+                    if (u === "throw")
+                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
+                    let d = reader.skip(wireType);
+                    if (u !== false)
+                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
+            }
+        }
+        return message;
+    }
+    internalBinaryWrite(message: GetSplitStatusResponse, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
+        /* bool in_progress = 1; */
+        if (message.inProgress !== false)
+            writer.tag(1, WireType.Varint).bool(message.inProgress);
+        /* string phase = 2; */
+        if (message.phase !== "")
+            writer.tag(2, WireType.LengthDelimited).string(message.phase);
+        /* string left_child_id = 3; */
+        if (message.leftChildId !== "")
+            writer.tag(3, WireType.LengthDelimited).string(message.leftChildId);
+        /* string right_child_id = 4; */
+        if (message.rightChildId !== "")
+            writer.tag(4, WireType.LengthDelimited).string(message.rightChildId);
+        /* string split_point = 5; */
+        if (message.splitPoint !== "")
+            writer.tag(5, WireType.LengthDelimited).string(message.splitPoint);
+        /* string initiator_node_id = 6; */
+        if (message.initiatorNodeId !== "")
+            writer.tag(6, WireType.LengthDelimited).string(message.initiatorNodeId);
+        /* int64 requested_at_ms = 7; */
+        if (message.requestedAtMs !== 0n)
+            writer.tag(7, WireType.Varint).int64(message.requestedAtMs);
+        let u = options.writeUnknownFields;
+        if (u !== false)
+            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
+        return writer;
+    }
+}
+/**
+ * @generated MessageType for protobuf message silo.v1.GetSplitStatusResponse
+ */
+export const GetSplitStatusResponse = new GetSplitStatusResponse$Type();
+// @generated message type with reflection information, may provide speed optimized methods
+class OwnedShardInfo$Type extends MessageType<OwnedShardInfo> {
+    constructor() {
+        super("silo.v1.OwnedShardInfo", [
+            { no: 1, name: "shard_id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
+            { no: 2, name: "total_jobs", kind: "scalar", T: 3 /*ScalarType.INT64*/, L: 0 /*LongType.BIGINT*/ },
+            { no: 3, name: "completed_jobs", kind: "scalar", T: 3 /*ScalarType.INT64*/, L: 0 /*LongType.BIGINT*/ },
+            { no: 4, name: "cleanup_status", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
+            { no: 5, name: "created_at_ms", kind: "scalar", T: 3 /*ScalarType.INT64*/, L: 0 /*LongType.BIGINT*/ },
+            { no: 6, name: "cleanup_completed_at_ms", kind: "scalar", T: 3 /*ScalarType.INT64*/, L: 0 /*LongType.BIGINT*/ }
+        ]);
+    }
+    create(value?: PartialMessage<OwnedShardInfo>): OwnedShardInfo {
+        const message = globalThis.Object.create((this.messagePrototype!));
+        message.shardId = "";
+        message.totalJobs = 0n;
+        message.completedJobs = 0n;
+        message.cleanupStatus = "";
+        message.createdAtMs = 0n;
+        message.cleanupCompletedAtMs = 0n;
+        if (value !== undefined)
+            reflectionMergePartial<OwnedShardInfo>(this, message, value);
+        return message;
+    }
+    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: OwnedShardInfo): OwnedShardInfo {
+        let message = target ?? this.create(), end = reader.pos + length;
+        while (reader.pos < end) {
+            let [fieldNo, wireType] = reader.tag();
+            switch (fieldNo) {
+                case /* string shard_id */ 1:
+                    message.shardId = reader.string();
+                    break;
+                case /* int64 total_jobs */ 2:
+                    message.totalJobs = reader.int64().toBigInt();
+                    break;
+                case /* int64 completed_jobs */ 3:
+                    message.completedJobs = reader.int64().toBigInt();
+                    break;
+                case /* string cleanup_status */ 4:
+                    message.cleanupStatus = reader.string();
+                    break;
+                case /* int64 created_at_ms */ 5:
+                    message.createdAtMs = reader.int64().toBigInt();
+                    break;
+                case /* int64 cleanup_completed_at_ms */ 6:
+                    message.cleanupCompletedAtMs = reader.int64().toBigInt();
+                    break;
+                default:
+                    let u = options.readUnknownField;
+                    if (u === "throw")
+                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
+                    let d = reader.skip(wireType);
+                    if (u !== false)
+                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
+            }
+        }
+        return message;
+    }
+    internalBinaryWrite(message: OwnedShardInfo, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
+        /* string shard_id = 1; */
+        if (message.shardId !== "")
+            writer.tag(1, WireType.LengthDelimited).string(message.shardId);
+        /* int64 total_jobs = 2; */
+        if (message.totalJobs !== 0n)
+            writer.tag(2, WireType.Varint).int64(message.totalJobs);
+        /* int64 completed_jobs = 3; */
+        if (message.completedJobs !== 0n)
+            writer.tag(3, WireType.Varint).int64(message.completedJobs);
+        /* string cleanup_status = 4; */
+        if (message.cleanupStatus !== "")
+            writer.tag(4, WireType.LengthDelimited).string(message.cleanupStatus);
+        /* int64 created_at_ms = 5; */
+        if (message.createdAtMs !== 0n)
+            writer.tag(5, WireType.Varint).int64(message.createdAtMs);
+        /* int64 cleanup_completed_at_ms = 6; */
+        if (message.cleanupCompletedAtMs !== 0n)
+            writer.tag(6, WireType.Varint).int64(message.cleanupCompletedAtMs);
+        let u = options.writeUnknownFields;
+        if (u !== false)
+            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
+        return writer;
+    }
+}
+/**
+ * @generated MessageType for protobuf message silo.v1.OwnedShardInfo
+ */
+export const OwnedShardInfo = new OwnedShardInfo$Type();
+// @generated message type with reflection information, may provide speed optimized methods
+class GetNodeInfoRequest$Type extends MessageType<GetNodeInfoRequest> {
+    constructor() {
+        super("silo.v1.GetNodeInfoRequest", []);
+    }
+    create(value?: PartialMessage<GetNodeInfoRequest>): GetNodeInfoRequest {
+        const message = globalThis.Object.create((this.messagePrototype!));
+        if (value !== undefined)
+            reflectionMergePartial<GetNodeInfoRequest>(this, message, value);
+        return message;
+    }
+    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: GetNodeInfoRequest): GetNodeInfoRequest {
+        let message = target ?? this.create(), end = reader.pos + length;
+        while (reader.pos < end) {
+            let [fieldNo, wireType] = reader.tag();
+            switch (fieldNo) {
+                default:
+                    let u = options.readUnknownField;
+                    if (u === "throw")
+                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
+                    let d = reader.skip(wireType);
+                    if (u !== false)
+                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
+            }
+        }
+        return message;
+    }
+    internalBinaryWrite(message: GetNodeInfoRequest, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
+        let u = options.writeUnknownFields;
+        if (u !== false)
+            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
+        return writer;
+    }
+}
+/**
+ * @generated MessageType for protobuf message silo.v1.GetNodeInfoRequest
+ */
+export const GetNodeInfoRequest = new GetNodeInfoRequest$Type();
+// @generated message type with reflection information, may provide speed optimized methods
+class GetNodeInfoResponse$Type extends MessageType<GetNodeInfoResponse> {
+    constructor() {
+        super("silo.v1.GetNodeInfoResponse", [
+            { no: 1, name: "node_id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
+            { no: 2, name: "owned_shards", kind: "message", repeat: 2 /*RepeatType.UNPACKED*/, T: () => OwnedShardInfo },
+            { no: 3, name: "placement_rings", kind: "scalar", repeat: 2 /*RepeatType.UNPACKED*/, T: 9 /*ScalarType.STRING*/ }
+        ]);
+    }
+    create(value?: PartialMessage<GetNodeInfoResponse>): GetNodeInfoResponse {
+        const message = globalThis.Object.create((this.messagePrototype!));
+        message.nodeId = "";
+        message.ownedShards = [];
+        message.placementRings = [];
+        if (value !== undefined)
+            reflectionMergePartial<GetNodeInfoResponse>(this, message, value);
+        return message;
+    }
+    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: GetNodeInfoResponse): GetNodeInfoResponse {
+        let message = target ?? this.create(), end = reader.pos + length;
+        while (reader.pos < end) {
+            let [fieldNo, wireType] = reader.tag();
+            switch (fieldNo) {
+                case /* string node_id */ 1:
+                    message.nodeId = reader.string();
+                    break;
+                case /* repeated silo.v1.OwnedShardInfo owned_shards */ 2:
+                    message.ownedShards.push(OwnedShardInfo.internalBinaryRead(reader, reader.uint32(), options));
+                    break;
+                case /* repeated string placement_rings */ 3:
+                    message.placementRings.push(reader.string());
+                    break;
+                default:
+                    let u = options.readUnknownField;
+                    if (u === "throw")
+                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
+                    let d = reader.skip(wireType);
+                    if (u !== false)
+                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
+            }
+        }
+        return message;
+    }
+    internalBinaryWrite(message: GetNodeInfoResponse, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
+        /* string node_id = 1; */
+        if (message.nodeId !== "")
+            writer.tag(1, WireType.LengthDelimited).string(message.nodeId);
+        /* repeated silo.v1.OwnedShardInfo owned_shards = 2; */
+        for (let i = 0; i < message.ownedShards.length; i++)
+            OwnedShardInfo.internalBinaryWrite(message.ownedShards[i], writer.tag(2, WireType.LengthDelimited).fork(), options).join();
+        /* repeated string placement_rings = 3; */
+        for (let i = 0; i < message.placementRings.length; i++)
+            writer.tag(3, WireType.LengthDelimited).string(message.placementRings[i]);
+        let u = options.writeUnknownFields;
+        if (u !== false)
+            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
+        return writer;
+    }
+}
+/**
+ * @generated MessageType for protobuf message silo.v1.GetNodeInfoResponse
+ */
+export const GetNodeInfoResponse = new GetNodeInfoResponse$Type();
+// @generated message type with reflection information, may provide speed optimized methods
+class ConfigureShardRequest$Type extends MessageType<ConfigureShardRequest> {
+    constructor() {
+        super("silo.v1.ConfigureShardRequest", [
+            { no: 1, name: "shard", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
+            { no: 2, name: "placement_ring", kind: "scalar", opt: true, T: 9 /*ScalarType.STRING*/ },
+            { no: 100, name: "tenant", kind: "scalar", opt: true, T: 9 /*ScalarType.STRING*/ }
+        ]);
+    }
+    create(value?: PartialMessage<ConfigureShardRequest>): ConfigureShardRequest {
+        const message = globalThis.Object.create((this.messagePrototype!));
+        message.shard = "";
+        if (value !== undefined)
+            reflectionMergePartial<ConfigureShardRequest>(this, message, value);
+        return message;
+    }
+    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: ConfigureShardRequest): ConfigureShardRequest {
+        let message = target ?? this.create(), end = reader.pos + length;
+        while (reader.pos < end) {
+            let [fieldNo, wireType] = reader.tag();
+            switch (fieldNo) {
+                case /* string shard */ 1:
+                    message.shard = reader.string();
+                    break;
+                case /* optional string placement_ring */ 2:
+                    message.placementRing = reader.string();
+                    break;
+                case /* optional string tenant */ 100:
+                    message.tenant = reader.string();
+                    break;
+                default:
+                    let u = options.readUnknownField;
+                    if (u === "throw")
+                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
+                    let d = reader.skip(wireType);
+                    if (u !== false)
+                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
+            }
+        }
+        return message;
+    }
+    internalBinaryWrite(message: ConfigureShardRequest, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
+        /* string shard = 1; */
+        if (message.shard !== "")
+            writer.tag(1, WireType.LengthDelimited).string(message.shard);
+        /* optional string placement_ring = 2; */
+        if (message.placementRing !== undefined)
+            writer.tag(2, WireType.LengthDelimited).string(message.placementRing);
+        /* optional string tenant = 100; */
+        if (message.tenant !== undefined)
+            writer.tag(100, WireType.LengthDelimited).string(message.tenant);
+        let u = options.writeUnknownFields;
+        if (u !== false)
+            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
+        return writer;
+    }
+}
+/**
+ * @generated MessageType for protobuf message silo.v1.ConfigureShardRequest
+ */
+export const ConfigureShardRequest = new ConfigureShardRequest$Type();
+// @generated message type with reflection information, may provide speed optimized methods
+class ConfigureShardResponse$Type extends MessageType<ConfigureShardResponse> {
+    constructor() {
+        super("silo.v1.ConfigureShardResponse", [
+            { no: 1, name: "previous_ring", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
+            { no: 2, name: "current_ring", kind: "scalar", T: 9 /*ScalarType.STRING*/ }
+        ]);
+    }
+    create(value?: PartialMessage<ConfigureShardResponse>): ConfigureShardResponse {
+        const message = globalThis.Object.create((this.messagePrototype!));
+        message.previousRing = "";
+        message.currentRing = "";
+        if (value !== undefined)
+            reflectionMergePartial<ConfigureShardResponse>(this, message, value);
+        return message;
+    }
+    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: ConfigureShardResponse): ConfigureShardResponse {
+        let message = target ?? this.create(), end = reader.pos + length;
+        while (reader.pos < end) {
+            let [fieldNo, wireType] = reader.tag();
+            switch (fieldNo) {
+                case /* string previous_ring */ 1:
+                    message.previousRing = reader.string();
+                    break;
+                case /* string current_ring */ 2:
+                    message.currentRing = reader.string();
+                    break;
+                default:
+                    let u = options.readUnknownField;
+                    if (u === "throw")
+                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
+                    let d = reader.skip(wireType);
+                    if (u !== false)
+                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
+            }
+        }
+        return message;
+    }
+    internalBinaryWrite(message: ConfigureShardResponse, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
+        /* string previous_ring = 1; */
+        if (message.previousRing !== "")
+            writer.tag(1, WireType.LengthDelimited).string(message.previousRing);
+        /* string current_ring = 2; */
+        if (message.currentRing !== "")
+            writer.tag(2, WireType.LengthDelimited).string(message.currentRing);
+        let u = options.writeUnknownFields;
+        if (u !== false)
+            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
+        return writer;
+    }
+}
+/**
+ * @generated MessageType for protobuf message silo.v1.ConfigureShardResponse
+ */
+export const ConfigureShardResponse = new ConfigureShardResponse$Type();
+// @generated message type with reflection information, may provide speed optimized methods
+class ImportAttempt$Type extends MessageType<ImportAttempt> {
+    constructor() {
+        super("silo.v1.ImportAttempt", [
+            { no: 1, name: "status", kind: "enum", T: () => ["silo.v1.AttemptStatus", AttemptStatus, "ATTEMPT_STATUS_"] },
+            { no: 2, name: "started_at_ms", kind: "scalar", T: 3 /*ScalarType.INT64*/, L: 0 /*LongType.BIGINT*/ },
+            { no: 3, name: "finished_at_ms", kind: "scalar", T: 3 /*ScalarType.INT64*/, L: 0 /*LongType.BIGINT*/ },
+            { no: 4, name: "result", kind: "message", T: () => SerializedBytes },
+            { no: 5, name: "error_code", kind: "scalar", opt: true, T: 9 /*ScalarType.STRING*/ },
+            { no: 6, name: "error_data", kind: "message", T: () => SerializedBytes }
+        ]);
+    }
+    create(value?: PartialMessage<ImportAttempt>): ImportAttempt {
+        const message = globalThis.Object.create((this.messagePrototype!));
+        message.status = 0;
+        message.startedAtMs = 0n;
+        message.finishedAtMs = 0n;
+        if (value !== undefined)
+            reflectionMergePartial<ImportAttempt>(this, message, value);
+        return message;
+    }
+    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: ImportAttempt): ImportAttempt {
+        let message = target ?? this.create(), end = reader.pos + length;
+        while (reader.pos < end) {
+            let [fieldNo, wireType] = reader.tag();
+            switch (fieldNo) {
+                case /* silo.v1.AttemptStatus status */ 1:
+                    message.status = reader.int32();
+                    break;
+                case /* int64 started_at_ms */ 2:
+                    message.startedAtMs = reader.int64().toBigInt();
+                    break;
+                case /* int64 finished_at_ms */ 3:
+                    message.finishedAtMs = reader.int64().toBigInt();
+                    break;
+                case /* optional silo.v1.SerializedBytes result */ 4:
+                    message.result = SerializedBytes.internalBinaryRead(reader, reader.uint32(), options, message.result);
+                    break;
+                case /* optional string error_code */ 5:
+                    message.errorCode = reader.string();
+                    break;
+                case /* optional silo.v1.SerializedBytes error_data */ 6:
+                    message.errorData = SerializedBytes.internalBinaryRead(reader, reader.uint32(), options, message.errorData);
+                    break;
+                default:
+                    let u = options.readUnknownField;
+                    if (u === "throw")
+                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
+                    let d = reader.skip(wireType);
+                    if (u !== false)
+                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
+            }
+        }
+        return message;
+    }
+    internalBinaryWrite(message: ImportAttempt, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
+        /* silo.v1.AttemptStatus status = 1; */
+        if (message.status !== 0)
+            writer.tag(1, WireType.Varint).int32(message.status);
+        /* int64 started_at_ms = 2; */
+        if (message.startedAtMs !== 0n)
+            writer.tag(2, WireType.Varint).int64(message.startedAtMs);
+        /* int64 finished_at_ms = 3; */
+        if (message.finishedAtMs !== 0n)
+            writer.tag(3, WireType.Varint).int64(message.finishedAtMs);
+        /* optional silo.v1.SerializedBytes result = 4; */
+        if (message.result)
+            SerializedBytes.internalBinaryWrite(message.result, writer.tag(4, WireType.LengthDelimited).fork(), options).join();
+        /* optional string error_code = 5; */
+        if (message.errorCode !== undefined)
+            writer.tag(5, WireType.LengthDelimited).string(message.errorCode);
+        /* optional silo.v1.SerializedBytes error_data = 6; */
+        if (message.errorData)
+            SerializedBytes.internalBinaryWrite(message.errorData, writer.tag(6, WireType.LengthDelimited).fork(), options).join();
+        let u = options.writeUnknownFields;
+        if (u !== false)
+            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
+        return writer;
+    }
+}
+/**
+ * @generated MessageType for protobuf message silo.v1.ImportAttempt
+ */
+export const ImportAttempt = new ImportAttempt$Type();
+// @generated message type with reflection information, may provide speed optimized methods
+class ImportJobRequest$Type extends MessageType<ImportJobRequest> {
+    constructor() {
+        super("silo.v1.ImportJobRequest", [
+            { no: 1, name: "shard", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
+            { no: 2, name: "id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
+            { no: 3, name: "priority", kind: "scalar", T: 13 /*ScalarType.UINT32*/ },
+            { no: 4, name: "enqueue_time_ms", kind: "scalar", T: 3 /*ScalarType.INT64*/, L: 0 /*LongType.BIGINT*/ },
+            { no: 5, name: "start_at_ms", kind: "scalar", T: 3 /*ScalarType.INT64*/, L: 0 /*LongType.BIGINT*/ },
+            { no: 6, name: "retry_policy", kind: "message", T: () => RetryPolicy },
+            { no: 7, name: "payload", kind: "message", T: () => SerializedBytes },
+            { no: 8, name: "limits", kind: "message", repeat: 2 /*RepeatType.UNPACKED*/, T: () => Limit },
+            { no: 9, name: "tenant", kind: "scalar", opt: true, T: 9 /*ScalarType.STRING*/ },
+            { no: 10, name: "metadata", kind: "map", K: 9 /*ScalarType.STRING*/, V: { kind: "scalar", T: 9 /*ScalarType.STRING*/ } },
+            { no: 11, name: "task_group", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
+            { no: 12, name: "attempts", kind: "message", repeat: 2 /*RepeatType.UNPACKED*/, T: () => ImportAttempt }
+        ]);
+    }
+    create(value?: PartialMessage<ImportJobRequest>): ImportJobRequest {
+        const message = globalThis.Object.create((this.messagePrototype!));
+        message.shard = "";
+        message.id = "";
+        message.priority = 0;
+        message.enqueueTimeMs = 0n;
+        message.startAtMs = 0n;
+        message.limits = [];
+        message.metadata = {};
+        message.taskGroup = "";
+        message.attempts = [];
+        if (value !== undefined)
+            reflectionMergePartial<ImportJobRequest>(this, message, value);
+        return message;
+    }
+    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: ImportJobRequest): ImportJobRequest {
+        let message = target ?? this.create(), end = reader.pos + length;
+        while (reader.pos < end) {
+            let [fieldNo, wireType] = reader.tag();
+            switch (fieldNo) {
+                case /* string shard */ 1:
+                    message.shard = reader.string();
+                    break;
+                case /* string id */ 2:
+                    message.id = reader.string();
+                    break;
+                case /* uint32 priority */ 3:
+                    message.priority = reader.uint32();
+                    break;
+                case /* int64 enqueue_time_ms */ 4:
+                    message.enqueueTimeMs = reader.int64().toBigInt();
+                    break;
+                case /* int64 start_at_ms */ 5:
+                    message.startAtMs = reader.int64().toBigInt();
+                    break;
+                case /* optional silo.v1.RetryPolicy retry_policy */ 6:
+                    message.retryPolicy = RetryPolicy.internalBinaryRead(reader, reader.uint32(), options, message.retryPolicy);
+                    break;
+                case /* silo.v1.SerializedBytes payload */ 7:
+                    message.payload = SerializedBytes.internalBinaryRead(reader, reader.uint32(), options, message.payload);
+                    break;
+                case /* repeated silo.v1.Limit limits */ 8:
+                    message.limits.push(Limit.internalBinaryRead(reader, reader.uint32(), options));
+                    break;
+                case /* optional string tenant */ 9:
+                    message.tenant = reader.string();
+                    break;
+                case /* map<string, string> metadata */ 10:
+                    this.binaryReadMap10(message.metadata, reader, options);
+                    break;
+                case /* string task_group */ 11:
+                    message.taskGroup = reader.string();
+                    break;
+                case /* repeated silo.v1.ImportAttempt attempts */ 12:
+                    message.attempts.push(ImportAttempt.internalBinaryRead(reader, reader.uint32(), options));
+                    break;
+                default:
+                    let u = options.readUnknownField;
+                    if (u === "throw")
+                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
+                    let d = reader.skip(wireType);
+                    if (u !== false)
+                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
+            }
+        }
+        return message;
+    }
+    private binaryReadMap10(map: ImportJobRequest["metadata"], reader: IBinaryReader, options: BinaryReadOptions): void {
+        let len = reader.uint32(), end = reader.pos + len, key: keyof ImportJobRequest["metadata"] | undefined, val: ImportJobRequest["metadata"][any] | undefined;
+        while (reader.pos < end) {
+            let [fieldNo, wireType] = reader.tag();
+            switch (fieldNo) {
+                case 1:
+                    key = reader.string();
+                    break;
+                case 2:
+                    val = reader.string();
+                    break;
+                default: throw new globalThis.Error("unknown map entry field for silo.v1.ImportJobRequest.metadata");
+            }
+        }
+        map[key ?? ""] = val ?? "";
+    }
+    internalBinaryWrite(message: ImportJobRequest, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
+        /* string shard = 1; */
+        if (message.shard !== "")
+            writer.tag(1, WireType.LengthDelimited).string(message.shard);
+        /* string id = 2; */
+        if (message.id !== "")
+            writer.tag(2, WireType.LengthDelimited).string(message.id);
+        /* uint32 priority = 3; */
+        if (message.priority !== 0)
+            writer.tag(3, WireType.Varint).uint32(message.priority);
+        /* int64 enqueue_time_ms = 4; */
+        if (message.enqueueTimeMs !== 0n)
+            writer.tag(4, WireType.Varint).int64(message.enqueueTimeMs);
+        /* int64 start_at_ms = 5; */
+        if (message.startAtMs !== 0n)
+            writer.tag(5, WireType.Varint).int64(message.startAtMs);
+        /* optional silo.v1.RetryPolicy retry_policy = 6; */
+        if (message.retryPolicy)
+            RetryPolicy.internalBinaryWrite(message.retryPolicy, writer.tag(6, WireType.LengthDelimited).fork(), options).join();
+        /* silo.v1.SerializedBytes payload = 7; */
+        if (message.payload)
+            SerializedBytes.internalBinaryWrite(message.payload, writer.tag(7, WireType.LengthDelimited).fork(), options).join();
+        /* repeated silo.v1.Limit limits = 8; */
+        for (let i = 0; i < message.limits.length; i++)
+            Limit.internalBinaryWrite(message.limits[i], writer.tag(8, WireType.LengthDelimited).fork(), options).join();
+        /* optional string tenant = 9; */
+        if (message.tenant !== undefined)
+            writer.tag(9, WireType.LengthDelimited).string(message.tenant);
+        /* map<string, string> metadata = 10; */
+        for (let k of globalThis.Object.keys(message.metadata))
+            writer.tag(10, WireType.LengthDelimited).fork().tag(1, WireType.LengthDelimited).string(k).tag(2, WireType.LengthDelimited).string(message.metadata[k]).join();
+        /* string task_group = 11; */
+        if (message.taskGroup !== "")
+            writer.tag(11, WireType.LengthDelimited).string(message.taskGroup);
+        /* repeated silo.v1.ImportAttempt attempts = 12; */
+        for (let i = 0; i < message.attempts.length; i++)
+            ImportAttempt.internalBinaryWrite(message.attempts[i], writer.tag(12, WireType.LengthDelimited).fork(), options).join();
+        let u = options.writeUnknownFields;
+        if (u !== false)
+            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
+        return writer;
+    }
+}
+/**
+ * @generated MessageType for protobuf message silo.v1.ImportJobRequest
+ */
+export const ImportJobRequest = new ImportJobRequest$Type();
+// @generated message type with reflection information, may provide speed optimized methods
+class ImportJobsRequest$Type extends MessageType<ImportJobsRequest> {
+    constructor() {
+        super("silo.v1.ImportJobsRequest", [
+            { no: 1, name: "jobs", kind: "message", repeat: 2 /*RepeatType.UNPACKED*/, T: () => ImportJobRequest }
+        ]);
+    }
+    create(value?: PartialMessage<ImportJobsRequest>): ImportJobsRequest {
+        const message = globalThis.Object.create((this.messagePrototype!));
+        message.jobs = [];
+        if (value !== undefined)
+            reflectionMergePartial<ImportJobsRequest>(this, message, value);
+        return message;
+    }
+    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: ImportJobsRequest): ImportJobsRequest {
+        let message = target ?? this.create(), end = reader.pos + length;
+        while (reader.pos < end) {
+            let [fieldNo, wireType] = reader.tag();
+            switch (fieldNo) {
+                case /* repeated silo.v1.ImportJobRequest jobs */ 1:
+                    message.jobs.push(ImportJobRequest.internalBinaryRead(reader, reader.uint32(), options));
+                    break;
+                default:
+                    let u = options.readUnknownField;
+                    if (u === "throw")
+                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
+                    let d = reader.skip(wireType);
+                    if (u !== false)
+                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
+            }
+        }
+        return message;
+    }
+    internalBinaryWrite(message: ImportJobsRequest, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
+        /* repeated silo.v1.ImportJobRequest jobs = 1; */
+        for (let i = 0; i < message.jobs.length; i++)
+            ImportJobRequest.internalBinaryWrite(message.jobs[i], writer.tag(1, WireType.LengthDelimited).fork(), options).join();
+        let u = options.writeUnknownFields;
+        if (u !== false)
+            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
+        return writer;
+    }
+}
+/**
+ * @generated MessageType for protobuf message silo.v1.ImportJobsRequest
+ */
+export const ImportJobsRequest = new ImportJobsRequest$Type();
+// @generated message type with reflection information, may provide speed optimized methods
+class ImportJobResult$Type extends MessageType<ImportJobResult> {
+    constructor() {
+        super("silo.v1.ImportJobResult", [
+            { no: 1, name: "id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
+            { no: 2, name: "success", kind: "scalar", T: 8 /*ScalarType.BOOL*/ },
+            { no: 3, name: "error", kind: "scalar", opt: true, T: 9 /*ScalarType.STRING*/ },
+            { no: 4, name: "status", kind: "enum", T: () => ["silo.v1.JobStatus", JobStatus, "JOB_STATUS_"] }
+        ]);
+    }
+    create(value?: PartialMessage<ImportJobResult>): ImportJobResult {
+        const message = globalThis.Object.create((this.messagePrototype!));
+        message.id = "";
+        message.success = false;
+        message.status = 0;
+        if (value !== undefined)
+            reflectionMergePartial<ImportJobResult>(this, message, value);
+        return message;
+    }
+    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: ImportJobResult): ImportJobResult {
+        let message = target ?? this.create(), end = reader.pos + length;
+        while (reader.pos < end) {
+            let [fieldNo, wireType] = reader.tag();
+            switch (fieldNo) {
+                case /* string id */ 1:
+                    message.id = reader.string();
+                    break;
+                case /* bool success */ 2:
+                    message.success = reader.bool();
+                    break;
+                case /* optional string error */ 3:
+                    message.error = reader.string();
+                    break;
+                case /* silo.v1.JobStatus status */ 4:
+                    message.status = reader.int32();
+                    break;
+                default:
+                    let u = options.readUnknownField;
+                    if (u === "throw")
+                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
+                    let d = reader.skip(wireType);
+                    if (u !== false)
+                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
+            }
+        }
+        return message;
+    }
+    internalBinaryWrite(message: ImportJobResult, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
+        /* string id = 1; */
+        if (message.id !== "")
+            writer.tag(1, WireType.LengthDelimited).string(message.id);
+        /* bool success = 2; */
+        if (message.success !== false)
+            writer.tag(2, WireType.Varint).bool(message.success);
+        /* optional string error = 3; */
+        if (message.error !== undefined)
+            writer.tag(3, WireType.LengthDelimited).string(message.error);
+        /* silo.v1.JobStatus status = 4; */
+        if (message.status !== 0)
+            writer.tag(4, WireType.Varint).int32(message.status);
+        let u = options.writeUnknownFields;
+        if (u !== false)
+            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
+        return writer;
+    }
+}
+/**
+ * @generated MessageType for protobuf message silo.v1.ImportJobResult
+ */
+export const ImportJobResult = new ImportJobResult$Type();
+// @generated message type with reflection information, may provide speed optimized methods
+class ImportJobsResponse$Type extends MessageType<ImportJobsResponse> {
+    constructor() {
+        super("silo.v1.ImportJobsResponse", [
+            { no: 1, name: "results", kind: "message", repeat: 2 /*RepeatType.UNPACKED*/, T: () => ImportJobResult }
+        ]);
+    }
+    create(value?: PartialMessage<ImportJobsResponse>): ImportJobsResponse {
+        const message = globalThis.Object.create((this.messagePrototype!));
+        message.results = [];
+        if (value !== undefined)
+            reflectionMergePartial<ImportJobsResponse>(this, message, value);
+        return message;
+    }
+    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: ImportJobsResponse): ImportJobsResponse {
+        let message = target ?? this.create(), end = reader.pos + length;
+        while (reader.pos < end) {
+            let [fieldNo, wireType] = reader.tag();
+            switch (fieldNo) {
+                case /* repeated silo.v1.ImportJobResult results */ 1:
+                    message.results.push(ImportJobResult.internalBinaryRead(reader, reader.uint32(), options));
+                    break;
+                default:
+                    let u = options.readUnknownField;
+                    if (u === "throw")
+                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
+                    let d = reader.skip(wireType);
+                    if (u !== false)
+                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
+            }
+        }
+        return message;
+    }
+    internalBinaryWrite(message: ImportJobsResponse, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
+        /* repeated silo.v1.ImportJobResult results = 1; */
+        for (let i = 0; i < message.results.length; i++)
+            ImportJobResult.internalBinaryWrite(message.results[i], writer.tag(1, WireType.LengthDelimited).fork(), options).join();
+        let u = options.writeUnknownFields;
+        if (u !== false)
+            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
+        return writer;
+    }
+}
+/**
+ * @generated MessageType for protobuf message silo.v1.ImportJobsResponse
+ */
+export const ImportJobsResponse = new ImportJobsResponse$Type();
+// @generated message type with reflection information, may provide speed optimized methods
+class ForceReleaseShardRequest$Type extends MessageType<ForceReleaseShardRequest> {
+    constructor() {
+        super("silo.v1.ForceReleaseShardRequest", [
+            { no: 1, name: "shard", kind: "scalar", T: 9 /*ScalarType.STRING*/ }
+        ]);
+    }
+    create(value?: PartialMessage<ForceReleaseShardRequest>): ForceReleaseShardRequest {
+        const message = globalThis.Object.create((this.messagePrototype!));
+        message.shard = "";
+        if (value !== undefined)
+            reflectionMergePartial<ForceReleaseShardRequest>(this, message, value);
+        return message;
+    }
+    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: ForceReleaseShardRequest): ForceReleaseShardRequest {
+        let message = target ?? this.create(), end = reader.pos + length;
+        while (reader.pos < end) {
+            let [fieldNo, wireType] = reader.tag();
+            switch (fieldNo) {
+                case /* string shard */ 1:
+                    message.shard = reader.string();
+                    break;
+                default:
+                    let u = options.readUnknownField;
+                    if (u === "throw")
+                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
+                    let d = reader.skip(wireType);
+                    if (u !== false)
+                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
+            }
+        }
+        return message;
+    }
+    internalBinaryWrite(message: ForceReleaseShardRequest, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
+        /* string shard = 1; */
+        if (message.shard !== "")
+            writer.tag(1, WireType.LengthDelimited).string(message.shard);
+        let u = options.writeUnknownFields;
+        if (u !== false)
+            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
+        return writer;
+    }
+}
+/**
+ * @generated MessageType for protobuf message silo.v1.ForceReleaseShardRequest
+ */
+export const ForceReleaseShardRequest = new ForceReleaseShardRequest$Type();
+// @generated message type with reflection information, may provide speed optimized methods
+class ForceReleaseShardResponse$Type extends MessageType<ForceReleaseShardResponse> {
+    constructor() {
+        super("silo.v1.ForceReleaseShardResponse", [
+            { no: 1, name: "released", kind: "scalar", T: 8 /*ScalarType.BOOL*/ }
+        ]);
+    }
+    create(value?: PartialMessage<ForceReleaseShardResponse>): ForceReleaseShardResponse {
+        const message = globalThis.Object.create((this.messagePrototype!));
+        message.released = false;
+        if (value !== undefined)
+            reflectionMergePartial<ForceReleaseShardResponse>(this, message, value);
+        return message;
+    }
+    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: ForceReleaseShardResponse): ForceReleaseShardResponse {
+        let message = target ?? this.create(), end = reader.pos + length;
+        while (reader.pos < end) {
+            let [fieldNo, wireType] = reader.tag();
+            switch (fieldNo) {
+                case /* bool released */ 1:
+                    message.released = reader.bool();
+                    break;
+                default:
+                    let u = options.readUnknownField;
+                    if (u === "throw")
+                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
+                    let d = reader.skip(wireType);
+                    if (u !== false)
+                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
+            }
+        }
+        return message;
+    }
+    internalBinaryWrite(message: ForceReleaseShardResponse, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
+        /* bool released = 1; */
+        if (message.released !== false)
+            writer.tag(1, WireType.Varint).bool(message.released);
+        let u = options.writeUnknownFields;
+        if (u !== false)
+            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
+        return writer;
+    }
+}
+/**
+ * @generated MessageType for protobuf message silo.v1.ForceReleaseShardResponse
+ */
+export const ForceReleaseShardResponse = new ForceReleaseShardResponse$Type();
+// @generated message type with reflection information, may provide speed optimized methods
+class CompactShardRequest$Type extends MessageType<CompactShardRequest> {
+    constructor() {
+        super("silo.v1.CompactShardRequest", [
+            { no: 1, name: "shard", kind: "scalar", T: 9 /*ScalarType.STRING*/ }
+        ]);
+    }
+    create(value?: PartialMessage<CompactShardRequest>): CompactShardRequest {
+        const message = globalThis.Object.create((this.messagePrototype!));
+        message.shard = "";
+        if (value !== undefined)
+            reflectionMergePartial<CompactShardRequest>(this, message, value);
+        return message;
+    }
+    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: CompactShardRequest): CompactShardRequest {
+        let message = target ?? this.create(), end = reader.pos + length;
+        while (reader.pos < end) {
+            let [fieldNo, wireType] = reader.tag();
+            switch (fieldNo) {
+                case /* string shard */ 1:
+                    message.shard = reader.string();
+                    break;
+                default:
+                    let u = options.readUnknownField;
+                    if (u === "throw")
+                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
+                    let d = reader.skip(wireType);
+                    if (u !== false)
+                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
+            }
+        }
+        return message;
+    }
+    internalBinaryWrite(message: CompactShardRequest, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
+        /* string shard = 1; */
+        if (message.shard !== "")
+            writer.tag(1, WireType.LengthDelimited).string(message.shard);
+        let u = options.writeUnknownFields;
+        if (u !== false)
+            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
+        return writer;
+    }
+}
+/**
+ * @generated MessageType for protobuf message silo.v1.CompactShardRequest
+ */
+export const CompactShardRequest = new CompactShardRequest$Type();
+// @generated message type with reflection information, may provide speed optimized methods
+class CompactShardResponse$Type extends MessageType<CompactShardResponse> {
+    constructor() {
+        super("silo.v1.CompactShardResponse", [
+            { no: 1, name: "status", kind: "scalar", T: 9 /*ScalarType.STRING*/ }
+        ]);
+    }
+    create(value?: PartialMessage<CompactShardResponse>): CompactShardResponse {
+        const message = globalThis.Object.create((this.messagePrototype!));
+        message.status = "";
+        if (value !== undefined)
+            reflectionMergePartial<CompactShardResponse>(this, message, value);
+        return message;
+    }
+    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: CompactShardResponse): CompactShardResponse {
+        let message = target ?? this.create(), end = reader.pos + length;
+        while (reader.pos < end) {
+            let [fieldNo, wireType] = reader.tag();
+            switch (fieldNo) {
+                case /* string status */ 1:
+                    message.status = reader.string();
+                    break;
+                default:
+                    let u = options.readUnknownField;
+                    if (u === "throw")
+                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
+                    let d = reader.skip(wireType);
+                    if (u !== false)
+                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
+            }
+        }
+        return message;
+    }
+    internalBinaryWrite(message: CompactShardResponse, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
+        /* string status = 1; */
+        if (message.status !== "")
+            writer.tag(1, WireType.LengthDelimited).string(message.status);
+        let u = options.writeUnknownFields;
+        if (u !== false)
+            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
+        return writer;
+    }
+}
+/**
+ * @generated MessageType for protobuf message silo.v1.CompactShardResponse
+ */
+export const CompactShardResponse = new CompactShardResponse$Type();
+// @generated message type with reflection information, may provide speed optimized methods
+class ReadLsmStateRequest$Type extends MessageType<ReadLsmStateRequest> {
+    constructor() {
+        super("silo.v1.ReadLsmStateRequest", [
+            { no: 1, name: "shard", kind: "scalar", T: 9 /*ScalarType.STRING*/ }
+        ]);
+    }
+    create(value?: PartialMessage<ReadLsmStateRequest>): ReadLsmStateRequest {
+        const message = globalThis.Object.create((this.messagePrototype!));
+        message.shard = "";
+        if (value !== undefined)
+            reflectionMergePartial<ReadLsmStateRequest>(this, message, value);
+        return message;
+    }
+    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: ReadLsmStateRequest): ReadLsmStateRequest {
+        let message = target ?? this.create(), end = reader.pos + length;
+        while (reader.pos < end) {
+            let [fieldNo, wireType] = reader.tag();
+            switch (fieldNo) {
+                case /* string shard */ 1:
+                    message.shard = reader.string();
+                    break;
+                default:
+                    let u = options.readUnknownField;
+                    if (u === "throw")
+                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
+                    let d = reader.skip(wireType);
+                    if (u !== false)
+                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
+            }
+        }
+        return message;
+    }
+    internalBinaryWrite(message: ReadLsmStateRequest, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
+        /* string shard = 1; */
+        if (message.shard !== "")
+            writer.tag(1, WireType.LengthDelimited).string(message.shard);
+        let u = options.writeUnknownFields;
+        if (u !== false)
+            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
+        return writer;
+    }
+}
+/**
+ * @generated MessageType for protobuf message silo.v1.ReadLsmStateRequest
+ */
+export const ReadLsmStateRequest = new ReadLsmStateRequest$Type();
+// @generated message type with reflection information, may provide speed optimized methods
+class LsmSstInfo$Type extends MessageType<LsmSstInfo> {
+    constructor() {
+        super("silo.v1.LsmSstInfo", [
+            { no: 1, name: "id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
+            { no: 2, name: "estimated_size", kind: "scalar", T: 4 /*ScalarType.UINT64*/, L: 0 /*LongType.BIGINT*/ }
+        ]);
+    }
+    create(value?: PartialMessage<LsmSstInfo>): LsmSstInfo {
+        const message = globalThis.Object.create((this.messagePrototype!));
+        message.id = "";
+        message.estimatedSize = 0n;
+        if (value !== undefined)
+            reflectionMergePartial<LsmSstInfo>(this, message, value);
+        return message;
+    }
+    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: LsmSstInfo): LsmSstInfo {
+        let message = target ?? this.create(), end = reader.pos + length;
+        while (reader.pos < end) {
+            let [fieldNo, wireType] = reader.tag();
+            switch (fieldNo) {
+                case /* string id */ 1:
+                    message.id = reader.string();
+                    break;
+                case /* uint64 estimated_size */ 2:
+                    message.estimatedSize = reader.uint64().toBigInt();
+                    break;
+                default:
+                    let u = options.readUnknownField;
+                    if (u === "throw")
+                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
+                    let d = reader.skip(wireType);
+                    if (u !== false)
+                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
+            }
+        }
+        return message;
+    }
+    internalBinaryWrite(message: LsmSstInfo, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
+        /* string id = 1; */
+        if (message.id !== "")
+            writer.tag(1, WireType.LengthDelimited).string(message.id);
+        /* uint64 estimated_size = 2; */
+        if (message.estimatedSize !== 0n)
+            writer.tag(2, WireType.Varint).uint64(message.estimatedSize);
+        let u = options.writeUnknownFields;
+        if (u !== false)
+            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
+        return writer;
+    }
+}
+/**
+ * @generated MessageType for protobuf message silo.v1.LsmSstInfo
+ */
+export const LsmSstInfo = new LsmSstInfo$Type();
+// @generated message type with reflection information, may provide speed optimized methods
+class LsmSortedRunInfo$Type extends MessageType<LsmSortedRunInfo> {
+    constructor() {
+        super("silo.v1.LsmSortedRunInfo", [
+            { no: 1, name: "id", kind: "scalar", T: 13 /*ScalarType.UINT32*/ },
+            { no: 2, name: "sst_count", kind: "scalar", T: 13 /*ScalarType.UINT32*/ },
+            { no: 3, name: "estimated_size", kind: "scalar", T: 4 /*ScalarType.UINT64*/, L: 0 /*LongType.BIGINT*/ }
+        ]);
+    }
+    create(value?: PartialMessage<LsmSortedRunInfo>): LsmSortedRunInfo {
+        const message = globalThis.Object.create((this.messagePrototype!));
+        message.id = 0;
+        message.sstCount = 0;
+        message.estimatedSize = 0n;
+        if (value !== undefined)
+            reflectionMergePartial<LsmSortedRunInfo>(this, message, value);
+        return message;
+    }
+    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: LsmSortedRunInfo): LsmSortedRunInfo {
+        let message = target ?? this.create(), end = reader.pos + length;
+        while (reader.pos < end) {
+            let [fieldNo, wireType] = reader.tag();
+            switch (fieldNo) {
+                case /* uint32 id */ 1:
+                    message.id = reader.uint32();
+                    break;
+                case /* uint32 sst_count */ 2:
+                    message.sstCount = reader.uint32();
+                    break;
+                case /* uint64 estimated_size */ 3:
+                    message.estimatedSize = reader.uint64().toBigInt();
+                    break;
+                default:
+                    let u = options.readUnknownField;
+                    if (u === "throw")
+                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
+                    let d = reader.skip(wireType);
+                    if (u !== false)
+                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
+            }
+        }
+        return message;
+    }
+    internalBinaryWrite(message: LsmSortedRunInfo, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
+        /* uint32 id = 1; */
+        if (message.id !== 0)
+            writer.tag(1, WireType.Varint).uint32(message.id);
+        /* uint32 sst_count = 2; */
+        if (message.sstCount !== 0)
+            writer.tag(2, WireType.Varint).uint32(message.sstCount);
+        /* uint64 estimated_size = 3; */
+        if (message.estimatedSize !== 0n)
+            writer.tag(3, WireType.Varint).uint64(message.estimatedSize);
+        let u = options.writeUnknownFields;
+        if (u !== false)
+            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
+        return writer;
+    }
+}
+/**
+ * @generated MessageType for protobuf message silo.v1.LsmSortedRunInfo
+ */
+export const LsmSortedRunInfo = new LsmSortedRunInfo$Type();
+// @generated message type with reflection information, may provide speed optimized methods
+class ReadLsmStateResponse$Type extends MessageType<ReadLsmStateResponse> {
+    constructor() {
+        super("silo.v1.ReadLsmStateResponse", [
+            { no: 1, name: "l0_ssts", kind: "message", repeat: 2 /*RepeatType.UNPACKED*/, T: () => LsmSstInfo },
+            { no: 2, name: "sorted_runs", kind: "message", repeat: 2 /*RepeatType.UNPACKED*/, T: () => LsmSortedRunInfo },
+            { no: 3, name: "total_l0_size", kind: "scalar", T: 4 /*ScalarType.UINT64*/, L: 0 /*LongType.BIGINT*/ },
+            { no: 4, name: "total_sorted_run_size", kind: "scalar", T: 4 /*ScalarType.UINT64*/, L: 0 /*LongType.BIGINT*/ }
+        ]);
+    }
+    create(value?: PartialMessage<ReadLsmStateResponse>): ReadLsmStateResponse {
+        const message = globalThis.Object.create((this.messagePrototype!));
+        message.l0Ssts = [];
+        message.sortedRuns = [];
+        message.totalL0Size = 0n;
+        message.totalSortedRunSize = 0n;
+        if (value !== undefined)
+            reflectionMergePartial<ReadLsmStateResponse>(this, message, value);
+        return message;
+    }
+    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: ReadLsmStateResponse): ReadLsmStateResponse {
+        let message = target ?? this.create(), end = reader.pos + length;
+        while (reader.pos < end) {
+            let [fieldNo, wireType] = reader.tag();
+            switch (fieldNo) {
+                case /* repeated silo.v1.LsmSstInfo l0_ssts */ 1:
+                    message.l0Ssts.push(LsmSstInfo.internalBinaryRead(reader, reader.uint32(), options));
+                    break;
+                case /* repeated silo.v1.LsmSortedRunInfo sorted_runs */ 2:
+                    message.sortedRuns.push(LsmSortedRunInfo.internalBinaryRead(reader, reader.uint32(), options));
+                    break;
+                case /* uint64 total_l0_size */ 3:
+                    message.totalL0Size = reader.uint64().toBigInt();
+                    break;
+                case /* uint64 total_sorted_run_size */ 4:
+                    message.totalSortedRunSize = reader.uint64().toBigInt();
+                    break;
+                default:
+                    let u = options.readUnknownField;
+                    if (u === "throw")
+                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
+                    let d = reader.skip(wireType);
+                    if (u !== false)
+                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
+            }
+        }
+        return message;
+    }
+    internalBinaryWrite(message: ReadLsmStateResponse, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
+        /* repeated silo.v1.LsmSstInfo l0_ssts = 1; */
+        for (let i = 0; i < message.l0Ssts.length; i++)
+            LsmSstInfo.internalBinaryWrite(message.l0Ssts[i], writer.tag(1, WireType.LengthDelimited).fork(), options).join();
+        /* repeated silo.v1.LsmSortedRunInfo sorted_runs = 2; */
+        for (let i = 0; i < message.sortedRuns.length; i++)
+            LsmSortedRunInfo.internalBinaryWrite(message.sortedRuns[i], writer.tag(2, WireType.LengthDelimited).fork(), options).join();
+        /* uint64 total_l0_size = 3; */
+        if (message.totalL0Size !== 0n)
+            writer.tag(3, WireType.Varint).uint64(message.totalL0Size);
+        /* uint64 total_sorted_run_size = 4; */
+        if (message.totalSortedRunSize !== 0n)
+            writer.tag(4, WireType.Varint).uint64(message.totalSortedRunSize);
+        let u = options.writeUnknownFields;
+        if (u !== false)
+            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
+        return writer;
+    }
+}
+/**
+ * @generated MessageType for protobuf message silo.v1.ReadLsmStateResponse
+ */
+export const ReadLsmStateResponse = new ReadLsmStateResponse$Type();
+/**
+ * @generated ServiceType for protobuf service silo.v1.SiloAdmin
+ */
+export const SiloAdmin = new ServiceType("silo.v1.SiloAdmin", [
+    { name: "GetNodeInfo", options: {}, I: GetNodeInfoRequest, O: GetNodeInfoResponse },
+    { name: "CpuProfile", options: {}, I: CpuProfileRequest, O: CpuProfileResponse },
+    { name: "RequestSplit", options: {}, I: RequestSplitRequest, O: RequestSplitResponse },
+    { name: "GetSplitStatus", options: {}, I: GetSplitStatusRequest, O: GetSplitStatusResponse },
+    { name: "ConfigureShard", options: {}, I: ConfigureShardRequest, O: ConfigureShardResponse },
+    { name: "ImportJobs", options: {}, I: ImportJobsRequest, O: ImportJobsResponse },
+    { name: "ResetShards", options: {}, I: ResetShardsRequest, O: ResetShardsResponse },
+    { name: "ForceReleaseShard", options: {}, I: ForceReleaseShardRequest, O: ForceReleaseShardResponse },
+    { name: "CompactShard", options: {}, I: CompactShardRequest, O: CompactShardResponse },
+    { name: "ReadLsmState", options: {}, I: ReadLsmStateRequest, O: ReadLsmStateResponse }
+]);
